Possibilitate to POST a multipart form
The  -D <callable> can return a tuple for that case. The first element
of the tuple will be a dict to be added to the session headers (it will
include the multipart content-type with the boundary declaration). The
second element is the body, or an aiohttp Payload object (providing a
.write() method which generate and sends the body contents)

An example of this feature can be seen in the tests added.