Added --data-args option
This option allows to pass arguments to the function specified with
-D py: option

The arguments is simply a string that the function will receive in the
third parameter `options` as `options["data_args"]`. One possible usage
is to pass a JSON, but in fact the format is free as long as the
function knows how to parse it.

A test was added to illustrate the case in which the function builds a
FormData object from the keys and values received via --data-args