Added a new mode to inject requests per timeslot
This new mode does not simulate "closed loops". Each request is
triggered at some fixed instants, not depending on having received the
answer from previous requests.

A file contains the number of requests to inject per time-slot, and new
command line switches allow to specify the name of that file and the
length of the timeslot in seconds.

Every T seconds, a new thread is created, which runs a molotov scenario
with N workers to inject N requests in parallel. After all requests
receive the answer, the thread dies. It is possible to have several
concurrent threads, if the answers to previous timeslot are still
pending. A metric named "iodid_num_threads" is a gauge which is updated
(and notified to statsd server) each time a thread starts or ends.

TODO: add tests for this new functionality